Todos returned from server are rendered to screen.
The RECEIVED_TODOS action is picked up by a set of combined
reducers that take the list of todos returned from the api call
and populates an array of ids specific for that filter.

Also, a reducer updates the table of ids to todo objects.
When a new filter is chosen, we map the array of ids to the todo
objects in the getVisibleTodos() selector.

For the first request for a given filter, the screen takes a half
second to render. After this has happened once, the changes are
instant as we're caching the todos in our todo table.