Refactors reducers.
Index was replaced with the old todos file, as it was essentially
redundant. This has flattened our state object a little.

Reducers were pulled into seperate file, and a public api added to
each of those files, for accessing their respective data. This way
we don't expose details of the data structure to all the files that
use these reducers.

getVisibleTodos was updated to use these public apis.