Loading indicator is displayed
This is only displayed when there are no cached todos.

A new action has been created REQUEST_TODOS which is
fired before fetching todos.

The list reducer a combined reducer, containing the list
of ids and whether an update is currently being fetched
for this list:

{
  ids,
  isFetching,
}

The isFetching reducer sets the value to true when the
REQUEST_TODOS action is dispatched, and false when the
RECEIVE_TODOS action is dispatched.