Dispatch of REQUEST_TODOS automatic upon fetch
When fetching new todos, a REQUEST_TODOS action is
automatically dispatched. This removes the concern of
which actions are needed from the component, and pushes
that into the fetchTodos action creator.

We've also written our own thunk middleware that gives us
a powerful way to create action creators that need to
dispatch multiple actions during some async event.