FetchTodo fails early when request is in progress
We replace the self written thunk middleware with thunk-redux.
This middleware also provides a getState function into our thunk
as well as a dispatch.

We use these, in conjunction with getIsFetching to decide whether
we are currently making a request for a given list. If we are, we
return a resolved promise and bomb out early.

This fixes the issue we had where we could click the each FilterLink
and fire off a new request, even if one was already in progress.