Application now handles error with message & retry
A reducer has been added to handle the state of the error message.
This reducer is maintained on a per-todo-list basis.

A FetchError todo component has been created that is used by the
VisibleTodoList when an error has occured. This a presentational
component and relies on its user providing a message and a callback
for the retry button.

The VisibleTodoList component has an updated property map that specifies
dependency on the errorMessage part of the state. It uses a selector,
provided from the index reducer file, to get the error message from the
relevant list's state, using the filter state to work this out. This is
much like how the rest of our selectors are working.