Add todo now works again
Mock API methods were added to our API for add & toggle todos. The
action creator for addTodo is now an async action creator (and a thunk).
This action triggers an request to add a new todo to the 'database' and
dispatches an ADD_TODO_SUCCESS event upon success.

The byId & createList reducers have been updated to handle this action.
The byId simply returns the new state with this new ID mapped to the new
Todo.
The createList reducer behaves differently depending on the list
filter. If the filter is active or all, then the ID is added.